MemberDao, BoardDao(미완성)

+++ src/main/java/example/springboard/dao/MemberDao.java
@@ -3,7 +3,7 @@ package example.springboard.dao;
 import example.springboard.dto.Member;
 
 public interface MemberDao {
-    int addMember(Member member);
+    Long addMember(Member member);
     String getName(Long id);
     String getPassword(String name);
     Member getMember(String name);
